refactor: enemy inheritance redo
