+++ Assets/Scripts/Enemy/EnemyMovement.cs
@@ -1,7 +1,7 @@
 using Unity.VisualScripting;
 using UnityEngine;
 
-public class EnemyMovement : MultiStateObjectComponent
+public class EnemyMovement : MonoBehaviour
 {
     public float fallSpeed = 5.0f; // Falling speed
     public float moveSpeed = 2.0f; // Movement speed after landing
@@ -45,24 +45,5 @@ public class EnemyMovement : MultiStateObjectComponent
         // Check if the enemy is grounded by casting a ray downward
         RaycastHit2D hit = Physics2D.Raycast(transform.position, Vector2.down, 0.1f);
         return hit.collider != null;
-    }
-
-    public override void OnTimeStateChange(TimeState newTimeState)
-    {
-        switch (newTimeState)
-        {
-            case TimeState.Past:
-                gameObject.SetActive(false);
-                break;
-            
-            case TimeState.Present:
-                gameObject.SetActive(true);
-                break;
-        }
-    }
-
-    public override void OnInteract()
-    {
-        
     }
 }
