+++ Assets/Scripts/Enemy/MultiStateEnemy.cs
@@ -1,0 +1,19 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Serialization;
+
+public class MultiStateEnemy : MultiStateObjectComponent
+{
+    public bool shouldBePetTarget;
+
+    public override void OnTimeStateChange(TimeState newTimeState)
+    {
+        
+    }
+
+    public override void OnInteract()
+    {
+        
+    }
+}
